% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusnz.R
\docType{package}
\name{censusnz}
\alias{censusnz}
\title{censusnz: A package for accessing New Zealand census data}
\description{
The package provides two functions as an interface for retrieving variables and data from the db.censusnz package.
See ?censusnz::get_data or ?censusnz::get_variables for examples
}
