% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get New Zealand census data}
\usage{
get_data(geography = NULL, variables = NULL, year = 2018)
}
\arguments{
\item{geography}{A string of the geographic area to be selected. Must be one of SA1, SA2, LBA, DHB, TA, RC}

\item{variables}{A string or character vector of the variables to be selected. Can use get_variables() to examine available variables}

\item{year}{The year of data requested. Currently the only available year is 2018, which is the default}
}
\value{
The resulting dataframe for the requested geography, variables
}
\description{
Get New Zealand census data
}
\examples{
get_data("RC", c("maori_descent", "smoking_status"))
}
