% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_geography}
\alias{get_geography}
\title{Get geography dataframe given input string}
\usage{
get_geography(geography)
}
\arguments{
\item{geography}{A string of the geographic area to be selected. Must be one of SA1, SA2, LBA, DHB, TA, RC, WARD}
}
\value{
The resulting dataframe for the requested geography, unedited from db.censusnz
}
\description{
Get geography dataframe given input string
}
\examples{
get_geography("RC")
}
