% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusnz.R
\name{get_variables}
\alias{get_variables}
\title{Get NZ census variables}
\usage{
get_variables()
}
\value{
A table of available variables, which can be queried with get_data()
}
\description{
Get NZ census variables
}
\examples{
get_variables()
}
