% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot New Zealand census data}
\usage{
plot_data(
  geography = NULL,
  variables = NULL,
  regions = NULL,
  year = 2018,
  n = 6,
  exclude_total = TRUE,
  exclude_other = TRUE
)
}
\arguments{
\item{geography}{A string of the geographic area to be selected. Must be one of SA1, SA2, LBA, DHB, TA, RC, WARD}

\item{variables}{A string or character vector of the variables to be selected. Can use get_variables() to examine available variables}

\item{regions}{A string or vector of the regions to be selected}

\item{year}{The year of data requested. Currently the only available year is 2018, which is the default}

\item{n}{The number of areas to use as faceted plots}

\item{exclude_total}{Whether or not to exclude any data with name label containing the substring "Total"}

\item{exclude_other}{Whether or not to exclude the 'Other' column after data has been lumped}
}
\value{
A ggplot object containing a plot of each variable, for each geographic
region name, coloured by variable group, from 2018 NZ census data.
}
\description{
This function provides a drop-in replacement for `censusnz::get_data` that
  allows helpful graphs of the census data to be conveniently plotted. By
  default the function plots the 6 regions with the highest number of
  observations, excluding 'NZ Total' and 'Other' categories. Configuration of
  the number of regions as well as whether to exclude the total and other
  categories is possible.
}
\examples{
# Multiple variables with defaults
plot_data("RC", c("maori_descent", "smoking_status"))
# Customising the layout and exclusions
plot_data("RC",
          "maori_descent",
          n = 3,
          exclude_total = FALSE,
          exclude_other = FALSE)
}
