% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative.R
\name{plot_relative}
\alias{plot_relative}
\title{Plot relative levels for New Zealand census data}
\usage{
plot_relative(
  geography = NULL,
  variables = NULL,
  regions = NULL,
  year = 2018,
  n = 6,
  position = "fill",
  exclude_total = TRUE,
  exclude_other = TRUE
)
}
\arguments{
\item{geography}{A string of the geographic area to be selected. Must be one
of SA1, SA2, LBA, DHB, TA, RC, WARD}

\item{variables}{A string or vector of the variables to be selected. Can use
get_variables() to examine available variables}

\item{regions}{A string or vector of the regions to be selected}

\item{year}{The year of data requested. Currently the only available year is
2018, which is the default}

\item{n}{The number of areas to use as faceted plots}

\item{position}{The ggplot2 bar plot position, i.e. "fill" or "dodge"}

\item{exclude_total}{Whether or not to exclude any data with name label
containing the substring "Total"}

\item{exclude_other}{Whether or not to exclude the 'Other' column after data
has been lumped}
}
\value{
A ggplot object containing a plot of each variable, for each
  geographic region name, coloured by variable group, from 2018 NZ census
  data.
}
\description{
This function will plot the relative levels of certain variables for several
  named regions. By default this will take the top 6 regions, excluding total
  NZ or other summary groups. Several parameters can be configured including
  the position of the bars and the number of regions to plot.
}
\details{
Plotting proportions relative to other levels often provides much more
  intuitive or useful information as it allows quick comparison of a region
  to other regions or to the average, while avoiding scale issues created by
  few regions having large numbers of observations.
}
\examples{
plot_data("RC", c("maori_descent", "smoking_status"))
# Customising the layout and exclusions
plot_relative("RC",
          "maori_descent",
          n = 3,
          position = "dodge",
          exclude_total = FALSE,
          exclude_other = FALSE)
}
